refactor: hook 간소화
